+++ src/apps/weblib/js-api/vizzu.js
@@ -294,13 +294,20 @@ export default class Vizzu {
   }
 
   _processAnimParams(animTarget, animOptions) {
-    let targets = Array.isArray(animTarget) ? animTarget : [animTarget];
-    let opts = Array.isArray(animOptions) ? animOptions : [animOptions];
-    for (let i = 0; i < targets.length; i++) {
-      let target = targets[i];
-      let opt = opts.length === 1 ? opts[0] : opts[i];
-      this._setKeyframe(target, opt);
+    let anims = [];
+
+    if (Array.isArray(animTarget)) {
+      for (let target of animTarget)
+        if (target.target !== undefined)
+          anims.push({ target: target.target, options: target.options });
+        else anims.push({ target: target, options: undefined });
+    } else {
+      anims.push({ target: animTarget, options: animOptions });
     }
+
+    for (let anim of anims) this._setKeyframe(anim.target, anim.options);
+
+    this._setAnimation(animOptions);
   }
 
   _setKeyframe(animTarget, animOptions) {