+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Added
 
+- Multi keyframe animation support (one animation through multiple chart).
 - 'style' property returns the style object only filled with the user-set 
   values, all the values (returned by this property till 0.6.x) can be get
   using the new 'getComputedStyle()' method.